# perf(metainfo): optimized FromHTTPHeader mem alloc