```
goos: darwin
goarch: arm64
pkg: github.com/bytedance/gopkg/cloud/metainfo
cpu: Apple M2 Pro
                                     │  ./old.out  │             ./new.out              │
                                     │   sec/op    │   sec/op     vs base               │
FromHTTPHeader/FromHTTPHeader-0-12     238.8n ± 2%   70.49n ± 1%  -70.49% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-10-12    742.8n ± 1%   586.4n ± 3%  -21.06% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-20-12    1.353µ ± 0%   1.161µ ± 3%  -14.16% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-50-12    3.564µ ± 1%   2.906µ ± 2%  -18.46% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-100-12   7.199µ ± 1%   6.213µ ± 3%  -13.70% (p=0.002 n=6)
geomean                                2.253µ        1.872µ       -16.90%

                                     │  ./old.out   │              ./new.out              │
                                     │     B/op     │     B/op      vs base               │
FromHTTPHeader/FromHTTPHeader-0-12      1120.0 ± 0%      16.0 ± 0%  -98.57% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-10-12     1232.0 ± 0%     464.0 ± 0%  -62.34% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-20-12     1312.0 ± 0%     928.0 ± 0%  -29.27% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-50-12    3.766Ki ± 0%   2.267Ki ± 0%  -39.81% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-100-12   8.656Ki ± 0%   7.659Ki ± 0%  -11.52% (p=0.002 n=6)
geomean                                2.662Ki        1.634Ki       -38.63%

                                     │ ./old.out  │            ./new.out             │
                                     │ allocs/op  │ allocs/op   vs base              │
FromHTTPHeader/FromHTTPHeader-0-12     4.000 ± 0%   1.000 ± 0%  -75.0% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-10-12    13.00 ± 0%   12.00 ± 0%  -7.69% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-20-12    23.00 ± 0%   22.00 ± 0%  -4.35% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-50-12    55.00 ± 0%   52.00 ± 0%  -5.45% (p=0.002 n=6)
FromHTTPHeader/FromHTTPHeader-100-12   107.0 ± 0%   105.0 ± 0%  -1.87% (p=0.002 n=6)
geomean                                36.42        34.65       -4.86%
```